html/browsers/browsing-the-web/back-forward-cache/service-worker-clients.https.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="resources/helper.sub.js"></script>
<script src="/service-workers/service-worker/resources/test-helpers.sub.js"></script>
<script>

promise_test(async t => {
  const workerUrl = 'resources/clients-matchall-service-worker.js?pipe=header(Service-Worker-Allowed,../)';
  const registration = await service_worker_unregister_and_register(t, workerUrl, './');
  t.add_cleanup(_ => registration.unregister());
  await wait_for_state(t, registration.installing, 'activated');

  const pageA = new RemoteContext(token());
  const pageB = new RemoteContext(token());

  const urlA = location.origin + executorPath + pageA.context_id;
  const urlB = originCrossSite + executorPath + pageB.context_id;

  window.open(urlA, '_blank', 'noopener');

  await pageA.execute_script(waitForPageShow);

  const clients1 = await (await fetch('/get-clients-matchall')).json();

  await pageA.execute_script(
    (url) => prepareNavigation(() => {
      location.href = url;
    }),
    [urlB]);

  await pageB.execute_script(waitForPageShow);

  const clients2 = await (await fetch('/get-clients-matchall')).json();

  await pageB.execute_script(
    () => {
      prepareNavigation(() => { history.back(); });
    }
  );

  await pageA.execute_script(waitForPageShow);
  await assert_bfcached(pageA);

  const clients3 = await (await fetch('/get-clients-matchall')).json();

  assert_true(clients1.indexOf(urlA) >= 0, '1: Before navigation');
  assert_true(clients2.indexOf(urlA) < 0, '2: Before back navigation');
  assert_true(clients3.indexOf(urlA) >= 0, '3: After back navigation');
}, 'Serviceworker clients');
</script>
